// src/app/user/model/user.model.ts
export interface User {
  checked: any;
  activation: any;
  id: string;
  userName: string;
  password: string;
  isActive: number;
  email: string;
  mobile: string;
  firstName: string;
  middleName: string;
  lastName: string;
  dodId: string;
  photo: string;
  userType: string;
  fullName: string;
  squadronId: string;
  readinessScore: number;
  squadronName: string;
  hasAppointment?: boolean;
}